added icon + removed unused stuff
there was apparently already an icon i had added to the project but neglected to actually link into the manifest or the main application icon - that's in now.

there was a function i added in the Utility library way back that was for subdividing images - removed.
this was added back when i was attempting to do multithreaded interpolation - sadly, it looked terrible at the edges of the slices.