actual size zoom option
renamed ResetView() to ResetViewFit().
added ResetViewTrueSize() to center the image and set the scale to 1.0x (true image size).

added a button to the context menu for displaying the image at its actual size.

changed the background color menu init loop to use a list instead of a dictionary (makes more sense for a lot of reasons, not the least of which is a guaranteed iteration order).