better error messages
before, if the image was set but failed to load it would set image no null and subtitle to "File Not Found" - this gives no obvious sign of failure unless you read the top of the window.

changed SetImage() on failure to display a message box alerting the failure, complete with an error icon and error system sound.
additionally, it doesn't change the current image, so a failure does nothing aside from show the error message dialog.